requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-Login==0.6.3
PyMySQL==1.1.0
python-dotenv==1.0.0
Werkzeug==2.3.7
bcrypt==4.0.1
cryptography>=3.4.8
APScheduler==3.10.4
beautifulsoup4==4.12.2
requests==2.31.0
lxml==4.9.3
playwright==1.40.0
Pillow==10.1.0
numpy==1.24.3
scipy==1.11.4
pytz==2023.3